+++ packages/cad-viewer/src/locale/en/main.ts
@@ -90,6 +90,7 @@ export default {
     fontNotLoaded: 'Font "{fontName}" can not be loaded from "{url}"!',
     failedToGetAvaiableFonts: 'Failed to get avaiable fonts from "{url}"!',
     failedToOpenFile: 'Failed to open file "{fileName}"!',
-    fetchingDrawingFile: 'Fetching file ...'
+    fetchingDrawingFile: 'Fetching file ...',
+    unknownEntities: 'This drawing contains {count} unknown or unsupported entities! Those entities will not be shown.'
   }
 }