+++ packages/cad-viewer/src/locale/zh/main.ts
@@ -87,8 +87,9 @@ export default {
     loadingFonts: '正在加载字体...',
     loadingDwgConverter: '正在加载DWG转换器...',
     fontNotLoaded: '无法从"{url}"加载字体"{fontName}"！',
-    failedToGetAvaiableFonts: '无法从"{url}"获取可用的字体信息!',
-    failedToOpenFile: '无法打开文件"{fileName}"!',
-    fetchingDrawingFile: '正在加载图纸文件...'
+    failedToGetAvaiableFonts: '无法从"{url}"获取可用的字体信息！',
+    failedToOpenFile: '无法打开文件"{fileName}"！',
+    fetchingDrawingFile: '正在加载图纸文件...',
+    unknownEntities: '这张图纸中包含了{count}个未知或不支持的实体，这些实体将无法显示！'
   }
 }