+++ packages/svg-renderer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/svg-renderer
 
+## 0.0.19
+
+### Patch Changes
+
+- feat: show warning message if found some unknown entities after parsed drawing
+
 ## 0.0.18
 
 ### Patch Changes